# apps/core/spec/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "CoreX Service API",
    "description": "HTTP 状态码始终为 200；业务结果见响应体 body.code（200000=成功）。错误码规则见 /guide/error-codes.md。Exception.details 字段仅在开发环境返回。鉴权：JWT Bearer；Apifox 环境变量统一使用 {{token}}（来自 signin.data.token）。密码登录 / 注册 / 发 OTP 前须先 POST /api/v1/auth/captcha 获取 captchaKey（go-captcha 行为验证码）。",
    "contact": {
      "name": "CoreX Team",
      "email": "15638470820@163.com"
    },
    "license": {
      "name": "Proprietary"
    },
    "version": "0.1.0"
  },
  "servers": [
    {
      "url": "http://127.0.0.1:3000",
      "description": "development"
    },
    {
      "url": "https://api.example.com",
      "description": "production"
    }
  ],
  "paths": {
...
    "/api/v1/search/docs": {
      "get": {
        "tags": [
          "Search"
        ],
        "summary": "全文检索",
        "description": "需要 JWT。按 q 在 title/content 上 multi_match 检索。",
        "operationId": "search.toRead",
        "parameters": [
          {
            "name": "q",
            "in": "query",
            "description": "搜索关键词",
            "required": true,
            "schema": {
              "type": "string"
            },
            "example": "Redis"
          },
          {
            "name": "size",
            "in": "query",
            "description": "返回条数，默认 10，最大 100",
            "required": false,
            "schema": {
              "type": "integer",
              "format": "int64"
            },
            "example": 10
          }
        ],
        "responses": {
          "200": {
            "description": "未登录或参数/ES 错误",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Exception"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer_auth": []
          }
        ]
      },
      "post": {
        "tags": [
          "Search"
        ],
        "summary": "索引文档",
        "description": "需要 JWT。将文档写入 Elasticsearch（示范搜索投影，不以 ES 为权威库）。",
        "operationId": "search.toWrite",
        "requestBody": {
          "description": "待索引文档",
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/WriteP"
              },
              "example": {
                "content": "接入 Redis 与 Elasticsearch",
                "id": "doc-1",
                "title": "Rust 服务端"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "未登录或参数/ES 错误",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Exception"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer_auth": []
          }
        ]
      }
    },
...
    "/api/v1/tenants": {
      "get": {
        "tags": [
          "Tenant"
        ],
        "summary": "我所属的租户",
        "operationId": "tenant.toList",
        "responses": {
          "200": {
            "description": "未登录",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Exception"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer_auth": []
          }
        ]
      },
      "post": {
        "tags": [
          "Tenant"
        ],
        "summary": "创建租户",
        "operationId": "tenant.toWrite",
        "requestBody": {
          "description": "租户信息",
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/TenantWriteP"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "参数错误",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Exception"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer_auth": []
          }
        ]
      }
    },
    "/api/v1/tenants/{id}": {
      "get": {
        "tags": [
          "Tenant"
        ],
        "summary": "租户详情",
        "operationId": "tenant.toReadById",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "description": "租户 ID",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "未找到",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Exception"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer_auth": []
          }
        ]
      },
      "put": {
        "tags": [
          "Tenant"
        ],
        "summary": "更新租户",
        "operationId": "tenant.toUpdate",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "description": "租户 ID",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "description": "租户信息",
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/TenantUpdateP"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "权限不足",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Exception"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer_auth": []
          }
        ]
      },
      "delete": {
        "tags": [
          "Tenant"
        ],
        "summary": "删除租户",
        "operationId": "tenant.toRemove",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "description": "租户 ID",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "权限不足",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Exception"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer_auth": []
          }
        ]
      }
    },
...
    "/api/v1/tenants/{id}/orders": {
      "get": {
        "tags": [
          "Payment"
        ],
        "summary": "订单历史",
        "description": "按创建时间倒序返回最近 20 条；返回前会惰性关闭已超时的待支付订单。",
        "operationId": "payment.toList",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "description": "租户 ID",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "权限不足",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Exception"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer_auth": []
          }
        ]
      },
      "post": {
        "tags": [
          "Payment"
        ],
        "summary": "创建支付订单",
        "description": "**金额由服务端按档位定价**，请求体只接受档位与渠道；同一用户 + 档位 + 渠道的未过期待支付订单会直接复用（避免重复下单）。返回 `codeUrl` 由客户端渲染二维码，`orderExpiresAt` 为支付截止时间。",
        "operationId": "payment.toWrite",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "description": "租户 ID",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "description": "档位与支付渠道",
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/OrderP"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "档位不可购买 / 渠道未配置 / 上游下单失败",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Exception"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer_auth": []
          }
        ]
      }
    },
...
    "/api/v1/tenants/{id}/subscriptions": {
      "get": {
        "tags": [
          "Subscription"
        ],
        "summary": "租户订阅历史",
        "description": "返回该租户的全部订阅（含已取消 / 已到期），按创建时间倒序。\n\n个人租户的当日配额取决于此刻生效的订阅档位：有效订阅档位 > 免费档；团队租户与无租户身份走全局配额。\n\n需 JWT；调用者须为该租户 ACTIVE 成员（平台 ADMIN 可旁路）。\n\n`body.code`：200000 成功；非成员为权限类错误码。",
        "operationId": "subscription.toList",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "description": "租户 ID",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "非租户成员",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Exception"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer_auth": []
          }
        ]
      },
      "post": {
        "tags": [
          "Subscription"
        ],
        "summary": "开通 / 续订订阅",
        "description": "为**个人租户**开通或续订订阅（付费档位）。订阅**创建即生效**（`createdAt` 即生效时间）。\n\n同一时刻最多一条生效订阅：调用时会先把该租户仍为 ACTIVE 的旧订阅作废（未过期的 `expiresAt` 截断到当前时间），再插入新订阅。\n\n`plan` 必须存在于配置 `gateway.plan_daily_token_quota`；`expiresAt` 缺省或 null 表示永久有效。团队租户不可订阅（回落全局配额）。\n\n需 JWT；调用者须为租户 OWNER/ADMIN（平台 ADMIN 可旁路）。\n\n`body.code`：200000 成功；参数类错误码用于档位不存在、到期时间非法、非个人租户。",
        "operationId": "subscription.toWrite",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "description": "租户 ID",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "description": "订阅信息",
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/SubscribeP"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "参数错误 / 权限不足",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Exception"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer_auth": []
          }
        ]
      }
    },
    "/api/v1/tenants/{id}/subscriptions/{subscriptionID}": {
      "delete": {
        "tags": [
          "Subscription"
        ],
        "summary": "取消订阅",
        "description": "立即取消订阅：状态置为 CANCELED，并把有效期截断到当前时间；该租户配额随即回落免费档。\n\n需 JWT；调用者须为租户 OWNER/ADMIN（平台 ADMIN 可旁路）。\n\n`body.code`：200000 成功；订阅不存在或不属于该租户返回未找到类错误码。",
        "operationId": "subscription.toRemove",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "description": "租户 ID",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "subscriptionID",
            "in": "path",
            "description": "订阅 ID",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "未找到",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Exception"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer_auth": []
          }
        ]
      }
    },
...
    "/api/v1/users": {
      "get": {
        "tags": [
          "User"
        ],
        "summary": "获取用户列表",
        "description": "后台管理员接口，返回全部用户，需要 JWT 鉴权。",
        "operationId": "user.toRead",
        "responses": {
          "200": {
            "description": "未登录或权限不足",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Exception"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer_auth": []
          }
        ]
      },
      "post": {
        "tags": [
          "User"
        ],
        "summary": "创建用户",
        "description": "后台管理员接口，需要 JWT 鉴权。",
        "operationId": "user.toWrite",
        "requestBody": {
          "description": "新用户账号",
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/WriteP"
              },
              "example": {
                "password": "123456",
                "role": "USER",
                "username": "alice"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "用户名已存在等业务错误",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Exception"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer_auth": []
          }
        ]
      }
    },
    "/api/v1/users/{id}": {
      "get": {
        "tags": [
          "User"
        ],
        "summary": "获取用户详情",
        "operationId": "user.toReadById",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "description": "用户 UUID",
            "required": true,
            "schema": {
              "type": "string"
            },
            "example": "3c430c21-0891-43e1-bcd2-1a22eb4a5389"
          }
        ],
        "responses": {
          "200": {
            "description": "用户不存在（code=500101）",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Exception"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer_auth": []
          }
        ]
      },
      "put": {
        "tags": [
          "User"
        ],
        "summary": "更新用户",
        "operationId": "user.toUpdate",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "description": "用户 UUID",
            "required": true,
            "schema": {
              "type": "string"
            },
            "example": "3c430c21-0891-43e1-bcd2-1a22eb4a5389"
          }
        ],
        "requestBody": {
          "description": "可更新字段（均为可选）",
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateP"
              },
              "example": {
                "age": 30,
                "birthday": "1990-01-01",
                "email": "alice@example.com",
                "gender": "FEMALE",
                "phone": "13800138000",
                "role": "USER",
                "status": "ACTIVE",
                "username": "alice"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "用户不存在或参数错误",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Exception"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer_auth": []
          }
        ]
      },
      "delete": {
        "tags": [
          "User"
        ],
        "summary": "删除用户",
        "operationId": "user.toRemove",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "description": "用户 UUID",
            "required": true,
            "schema": {
              "type": "string"
            },
            "example": "3c430c21-0891-43e1-bcd2-1a22eb4a5389"
          }
        ],
        "responses": {
          "200": {
            "description": "用户不存在",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Exception"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer_auth": []
          }
        ]
      }
    }
  },
  "components": {
    "schemas": {
...
      "Exception": {
        "type": "object",
        "description": "异常响应结构",
        "required": [
          "code",
          "success",
          "msg",
          "timestamp"
        ],
        "properties": {
          "code": {
            "type": "integer",
            "format": "int32",
            "description": "业务错误码 (非0表示各种业务错误)"
          },
          "data": {
            "type": [
              "object",
              "null"
            ]
          },
          "details": {
            "type": [
              "object",
              "null"
            ],
            "description": "详细错误信息（仅非生产环境返回）"
          },
          "msg": {
            "type": "string"
          },
          "success": {
            "type": "boolean"
          },
          "timestamp": {
            "type": "integer",
            "format": "int64"
          }
        }
      },
...
      "OrderP": {
        "type": "object",
        "description": "创建订单请求。\n\n**只接受档位与渠道**：金额由服务端按 `pay.plans` 定价，客户端无法指定价格。",
        "required": [
          "plan",
          "channel"
        ],
        "properties": {
          "channel": {
            "type": "string",
            "description": "支付渠道：`WECHAT` / `ALIPAY`"
          },
          "plan": {
            "type": "string",
            "description": "档位名（须同时存在于 `pay.plans` 与 `gateway.planDailyTokenQuota`）"
          }
        }
      },
...
      "SubscribeP": {
        "type": "object",
        "description": "开通 / 续订请求。同一租户同一时刻最多一条生效订阅，重复调用即续订。",
        "required": [
          "plan"
        ],
        "properties": {
          "expiresAt": {
            "type": [
              "integer",
              "null"
            ],
            "format": "int64",
            "description": "到期时间（毫秒时间戳）；缺省或 null 为永久有效"
          },
          "plan": {
            "type": "string",
            "description": "档位名（须存在于 `gateway.plan_daily_token_quota`）"
          }
        }
      },
...
      "TenantUpdateP": {
        "type": "object",
        "properties": {
          "name": {
            "type": [
              "string",
              "null"
            ]
          },
          "status": {
            "type": [
              "string",
              "null"
            ]
          },
          "type": {
            "type": [
              "string",
              "null"
            ],
            "description": "PERSONAL / TEAM"
          }
        }
      },
      "TenantWriteP": {
        "type": "object",
        "required": [
          "name",
          "slug"
        ],
        "properties": {
          "name": {
            "type": "string"
          },
          "slug": {
            "type": "string",
            "description": "唯一标识（小写字母/数字/中划线）"
          },
          "type": {
            "type": [
              "string",
              "null"
            ],
            "description": "PERSONAL / TEAM；缺省 PERSONAL"
          }
        }
      },
      "UpdateP": {
        "type": "object",
        "description": "后台管理员更新用户（账号/密码/资料运维）\n用户自助修改个人信息请使用 PUT /api/v1/auth/profile",
        "properties": {
          "age": {
            "type": [
              "integer",
              "null"
            ],
            "format": "int32",
            "example": 30,
            "minimum": 0
          },
          "avatar": {
            "type": [
              "string",
              "null"
            ],
            "example": "550e8400-e29b-41d4-a716-446655440000"
          },
          "birthday": {
            "type": [
              "string",
              "null"
            ],
            "example": "1990-01-01"
          },
          "email": {
            "type": [
              "string",
              "null"
            ],
            "example": "alice@example.com"
          },
          "gender": {
            "type": [
              "string",
              "null"
            ],
            "example": "FEMALE"
          },
          "password": {
            "type": [
              "string",
              "null"
            ],
            "example": "123456"
          },
          "phone": {
            "type": [
              "string",
              "null"
            ],
            "example": "13800138000"
          },
          "role": {
            "type": [
              "string",
              "null"
            ],
            "example": "USER"
          },
          "status": {
            "type": [
              "string",
              "null"
            ],
            "example": "ACTIVE"
          },
          "username": {
            "type": [
              "string",
              "null"
            ],
            "example": "alice"
          }
        },
        "example": {
          "age": 30,
          "birthday": "1990-01-01",
          "email": "alice@example.com",
          "gender": "FEMALE",
          "phone": "13800138000",
          "role": "USER",
          "status": "ACTIVE",
          "username": "alice"
        }
      },
...
      "WriteP": {
        "type": "object",
        "required": [
          "title",
          "content"
        ],
        "properties": {
          "content": {
            "type": "string",
            "example": "接入 Redis 与 Elasticsearch"
          },
          "id": {
            "type": [
              "string",
              "null"
            ],
            "description": "文档 ID；省略则自动生成 UUID",
            "example": "doc-1"
          },
          "title": {
            "type": "string",
            "example": "Rust 服务端"
          }
        },
        "example": {
          "content": "接入 Redis 与 Elasticsearch",
          "id": "doc-1",
          "title": "Rust 服务端"
        }
      },
...
    },
    "securitySchemes": {
      "bearer_auth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "description": "登录 `POST /api/v1/auth/signin` 返回的 `data.token`。\n\n请求头：`Authorization: Bearer <token>`。\n\n**Apifox**：鉴权组件 Token 请填 `{{token}}`（勿用 `{{bearerToken}}`）；环境变量名统一为 `token`，可在登录接口后置操作写入。",
        "x-default": "{{token}}",
        "x-apifox-default": "{{token}}"
      }
    }
  },
...
}